Proof of concept oneshot channel #6 (#8)
Oneshot channel is about sending a value from one task to another within the same runtime, but you can also think about it about like of some kind execution switch between tasks. What happens is that sender's task is suspended, receiver's tasks awaits and executes something, then sender is awoken again. 

This implementation does not support concurrent oneshot channels, so it is not production ready, but for single channel there are a lot of tests.